Update IGRF when REALOP gets a new TLE
Problem: IGRF was updating every time we ran determination,
which is costly and unnecessary.

Solution: Rather than storing and calculating the relative
time since the last time we updated IGRF (which only FSW
can reliably tell us because CPU restart clears the memory
ADCS has access to), we can limit the amount of updating
we perform by only running it when we get a new TLE.
This is expected to happen weekly, and we will never
run determination with an un-updated IGRF because we cannot
run determination without a new(ish) TLE anyway.