Fix error guard in igrf_set_date_time
Problem: The guard that returns an error if the given time
is not appropriate to run IGRF only returns an error if
EVERY error condition is true.

Solution: Make the guard return an error if ANY of the error
conditions are true. All '&&' -> '||'.

Additionally, see https://github.com/risherlock/ADCS/issues/1